Add files via upload
Much more to work on 

dt determined by the speed of light, that should be further refinement.

particle creation is backward, should create particles that do not exist. the indexing in the creation is not optimal

zoom acts weird. 

 enable multi-particle processing. 

refinement of the forces or the dt, the circles against the analytical solution are too large. 


check the Lorentz transformation for the visualization of the fields, and maybe add the local point particles fields. 

energy loss momentum